@@ -27,7 +27,7 @@ import java.util.List;
 
 import static org.elasticsearch.rest.RestRequest.Method.POST;
 
-@ServerlessScope(Scope.PUBLIC)
+@ServerlessScope(Scope.INTERNAL)
 public class RestAnalyzeIndexDiskUsageAction extends BaseRestHandler {
 
     @Override
