@@ -13,25 +13,30 @@ import org.apache.lucene.util.BytesRef;
 import org.elasticsearch.Version;
 import org.elasticsearch.index.mapper.vectors.DenseVectorFieldMapper.ElementType;
 import org.elasticsearch.index.mapper.vectors.DenseVectorScriptDocValues;
+import org.elasticsearch.index.mapper.vectors.VectorEncoderDecoder;
 
 import java.io.IOException;
 
 public class BinaryDenseVectorDocValuesField extends DenseVectorDocValuesField {
 
-    protected final BinaryDocValues input;
-    protected final Version indexVersion;
-    protected final int dims;
-    protected BytesRef value;
+    private final BinaryDocValues input;
+    private final float[] vectorValue;
+    private final Version indexVersion;
+    private boolean decoded;
+    private final int dims;
+    private BytesRef value;
 
     public BinaryDenseVectorDocValuesField(BinaryDocValues input, String name, ElementType elementType, int dims, Version indexVersion) {
         super(name, elementType);
         this.input = input;
         this.indexVersion = indexVersion;
         this.dims = dims;
+        this.vectorValue = new float[dims];
     }
 
     @Override
     public void setNextDocId(int docId) throws IOException {
+        decoded = false;
         if (input.advanceExact(docId)) {
             value = input.binaryValue();
         } else {
@@ -54,8 +59,8 @@ public class BinaryDenseVectorDocValuesField extends DenseVectorDocValuesField {
         if (isEmpty()) {
             return DenseVector.EMPTY;
         }
-
-        return new BinaryDenseVector(value, dims, indexVersion);
+        decodeVectorIfNecessary();
+        return new BinaryDenseVector(vectorValue, value, dims, indexVersion);
     }
 
     @Override
@@ -63,11 +68,19 @@ public class BinaryDenseVectorDocValuesField extends DenseVectorDocValuesField {
         if (isEmpty()) {
             return defaultValue;
         }
-        return new BinaryDenseVector(value, dims, indexVersion);
+        decodeVectorIfNecessary();
+        return new BinaryDenseVector(vectorValue, value, dims, indexVersion);
     }
 
     @Override
     public DenseVector getInternal() {
         return get(null);
+    }
+
+    private void decodeVectorIfNecessary() {
+        if (decoded == false && value != null) {
+            VectorEncoderDecoder.decodeDenseVector(value, vectorValue);
+            decoded = true;
+        }
     }
 }
