@@ -24,9 +24,11 @@ import java.util.Collections;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
+import java.util.NoSuchElementException;
 import java.util.function.Predicate;
 import java.util.stream.Stream;
 
+import static java.util.Collections.emptyMap;
 import static java.util.stream.Collectors.toList;
 import static org.hamcrest.Matchers.equalTo;
 import static org.mockito.Mockito.mock;
@@ -55,7 +57,7 @@ public abstract class AbstractPercentilesTestCase<T extends InternalAggregation 
         for (int i = 0; i < numValues; ++i) {
             values[i] = randomDouble();
         }
-        return createTestInstance(name, metadata, keyed, format, percents, values);
+        return createTestInstance(name, metadata, keyed, format, percents, values, false);
     }
 
     protected abstract T createTestInstance(
@@ -64,7 +66,8 @@ public abstract class AbstractPercentilesTestCase<T extends InternalAggregation 
         boolean keyed,
         DocValueFormat format,
         double[] percents,
-        double[] values
+        double[] values,
+        boolean empty
     );
 
     protected abstract Class<? extends ParsedPercentiles> implementationClass();
@@ -104,7 +107,7 @@ public abstract class AbstractPercentilesTestCase<T extends InternalAggregation 
         boolean keyed = randomBoolean();
         DocValueFormat docValueFormat = randomNumericDocValueFormat();
 
-        T agg = createTestInstance("test", Collections.emptyMap(), keyed, docValueFormat, percents, new double[0]);
+        T agg = createTestInstance("test", Collections.emptyMap(), keyed, docValueFormat, percents, new double[0], false);
 
         for (Percentile percentile : agg) {
             Double value = percentile.getValue();
@@ -146,5 +149,27 @@ public abstract class AbstractPercentilesTestCase<T extends InternalAggregation 
         }
 
         assertThat(Strings.toString(builder), equalTo(expected));
+    }
+
+    public void testEmptyIterator() {
+        final double[] percents = randomPercents(false);
+
+        final Iterable<?> aggregation = createTestInstance(
+            "test",
+            emptyMap(),
+            false,
+            randomNumericDocValueFormat(),
+            percents,
+            new double[] {},
+            true
+        );
+
+        for (var ignored : aggregation) {
+            fail("empty expected");
+        }
+
+        final Iterator<?> it = aggregation.iterator();
+        assertFalse(it.hasNext());
+        expectThrows(NoSuchElementException.class, it::next);
     }
 }
