@@ -7,11 +7,13 @@
 package org.elasticsearch.xpack.core.search.action;
 
 import org.elasticsearch.ExceptionsHelper;
+import org.elasticsearch.TransportVersion;
 import org.elasticsearch.action.ActionResponse;
 import org.elasticsearch.action.search.SearchResponse;
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.io.stream.StreamOutput;
 import org.elasticsearch.common.xcontent.StatusToXContentObject;
+import org.elasticsearch.core.Nullable;
 import org.elasticsearch.rest.RestStatus;
 import org.elasticsearch.rest.action.RestActions;
 import org.elasticsearch.xcontent.XContentBuilder;
@@ -36,6 +38,10 @@ public class AsyncStatusResponse extends ActionResponse implements SearchStatusR
     private final int failedShards;
     private final RestStatus completionStatus;
 
+    // Non-null for cross cluster searches
+    @Nullable
+    private final SearchResponse.Clusters clusters;
+
     public AsyncStatusResponse(
         String id,
         boolean isRunning,
@@ -46,7 +52,8 @@ public class AsyncStatusResponse extends ActionResponse implements SearchStatusR
         int successfulShards,
         int skippedShards,
         int failedShards,
-        RestStatus completionStatus
+        RestStatus completionStatus,
+        SearchResponse.Clusters clusters
     ) {
         this.id = id;
         this.isRunning = isRunning;
@@ -58,6 +65,7 @@ public class AsyncStatusResponse extends ActionResponse implements SearchStatusR
         this.skippedShards = skippedShards;
         this.failedShards = failedShards;
         this.completionStatus = completionStatus;
+        this.clusters = clusters;
     }
 
     /**
@@ -76,6 +84,7 @@ public class AsyncStatusResponse extends ActionResponse implements SearchStatusR
         int successfulShards = 0;
         int skippedShards = 0;
         int failedShards = 0;
+        SearchResponse.Clusters clusters = null;
         RestStatus completionStatus = null;
         SearchResponse searchResponse = asyncSearchResponse.getSearchResponse();
         if (searchResponse != null) {
@@ -83,15 +92,18 @@ public class AsyncStatusResponse extends ActionResponse implements SearchStatusR
             successfulShards = searchResponse.getSuccessfulShards();
             skippedShards = searchResponse.getSkippedShards();
             failedShards = searchResponse.getFailedShards();
+            if (searchResponse.getClusters() != null && searchResponse.getClusters() != SearchResponse.Clusters.EMPTY) {
+                clusters = searchResponse.getClusters();
+            }
         }
         if (asyncSearchResponse.isRunning() == false) {
-            if (searchResponse != null) {
+            Exception failure = asyncSearchResponse.getFailure();
+            if (failure != null) {
+                completionStatus = ExceptionsHelper.status(ExceptionsHelper.unwrapCause(failure));
+            } else if (searchResponse != null) {
                 completionStatus = searchResponse.status();
             } else {
-                Exception failure = asyncSearchResponse.getFailure();
-                if (failure != null) {
-                    completionStatus = ExceptionsHelper.status(ExceptionsHelper.unwrapCause(failure));
-                }
+                throw new IllegalStateException("Unable to retrieve async_search status. No SearchResponse or Exception could be found.");
             }
         }
         return new AsyncStatusResponse(
@@ -104,7 +116,8 @@ public class AsyncStatusResponse extends ActionResponse implements SearchStatusR
             successfulShards,
             skippedShards,
             failedShards,
-            completionStatus
+            completionStatus,
+            clusters
         );
     }
 
@@ -119,6 +132,11 @@ public class AsyncStatusResponse extends ActionResponse implements SearchStatusR
         this.skippedShards = in.readVInt();
         this.failedShards = in.readVInt();
         this.completionStatus = (this.isRunning == false) ? RestStatus.readFrom(in) : null;
+        if (in.getTransportVersion().onOrAfter(TransportVersion.V_8_500_009)) {
+            this.clusters = in.readOptionalWriteable(SearchResponse.Clusters::new);
+        } else {
+            this.clusters = null;
+        }
     }
 
     @Override
@@ -134,6 +152,10 @@ public class AsyncStatusResponse extends ActionResponse implements SearchStatusR
         out.writeVInt(failedShards);
         if (isRunning == false) {
             RestStatus.writeTo(out, completionStatus);
+        }
+        if (out.getTransportVersion().onOrAfter(TransportVersion.V_8_500_009)) {
+            // optional since only CCS uses is; it is null for local-only searches
+            out.writeOptionalWriteable(clusters);
         }
     }
 
@@ -151,6 +173,9 @@ public class AsyncStatusResponse extends ActionResponse implements SearchStatusR
         builder.timeField("start_time_in_millis", "start_time", startTimeMillis);
         builder.timeField("expiration_time_in_millis", "expiration_time", expirationTimeMillis);
         RestActions.buildBroadcastShardsHeader(builder, params, totalShards, successfulShards, skippedShards, failedShards, null);
+        if (clusters != null) {
+            builder = clusters.toXContent(builder, null);
+        }
         if (isRunning == false) { // completion status information is only available for a completed search
             builder.field("completion_status", completionStatus.getStatus());
         }
@@ -172,7 +197,8 @@ public class AsyncStatusResponse extends ActionResponse implements SearchStatusR
             && successfulShards == other.successfulShards
             && skippedShards == other.skippedShards
             && failedShards == other.failedShards
-            && Objects.equals(completionStatus, other.completionStatus);
+            && Objects.equals(completionStatus, other.completionStatus)
+            && Objects.equals(clusters, other.clusters);
     }
 
     @Override
@@ -187,7 +213,8 @@ public class AsyncStatusResponse extends ActionResponse implements SearchStatusR
             successfulShards,
             skippedShards,
             failedShards,
-            completionStatus
+            completionStatus,
+            clusters
         );
     }
 
@@ -264,5 +291,13 @@ public class AsyncStatusResponse extends ActionResponse implements SearchStatusR
      */
     public RestStatus getCompletionStatus() {
         return completionStatus;
+    }
+
+    /**
+     * @return For CCS, clusters object that has information about the clustes being searched, such as total count
+     * successful count and skipped. Will be null for local-only searches.
+     */
+    public SearchResponse.Clusters getClusters() {
+        return clusters;
     }
 }
